requirements.txt
anthropic==0.97.0
chainlit==2.3.0
fastapi==0.115.14
faster-whisper==1.2.1
google-api-python-client==2.195.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-auth==2.50.0
httpx==0.28.1
imageio-ffmpeg==0.6.0
pydantic==2.13.3
python-telegram-bot==22.5
requests==2.32.5
starlette==0.41.3
uvicorn==0.39.0
